Refactor to expose only necessary APIs.
