+++ components/action-mgt/org.wso2.carbon.identity.action.execution/pom.xml
@@ -77,9 +77,13 @@
                         <Bundle-Name>${project.artifactId}</Bundle-Name>
                         <Private-Package>
                             org.wso2.carbon.identity.action.execution.internal,
+                            org.wso2.carbon.identity.action.execution.impl,
+                            org.wso2.carbon.identity.action.execution.util,
                         </Private-Package>
                         <Export-Package>
                             !org.wso2.carbon.identity.action.execution.internal,
+                            !org.wso2.carbon.identity.action.execution.impl,
+                            !org.wso2.carbon.identity.action.execution.util,
                             org.wso2.carbon.identity.action.execution.*;
                             version="${carbon.identity.package.export.version}"
                         </Export-Package>
